+++ src/components/computing-status/use-all-computing-status.ts
@@ -11,6 +11,7 @@ import {
     getSensitivityAnalysisRunningStatus,
     getShortCircuitAnalysisRunningStatus,
     getDynamicSimulationRunningStatus,
+    getVoltageInitRunningStatus,
     getLoadFlowRunningStatus,
     getNonEvacuatedEnergyRunningStatus,
 } from '../utils/running-status';
@@ -24,11 +25,11 @@ import {
     fetchOneBusShortCircuitAnalysisStatus,
     fetchShortCircuitAnalysisStatus,
 } from '../../services/study/short-circuit-analysis';
+import { fetchVoltageInitStatus } from '../../services/study/voltage-init';
 import { fetchLoadFlowStatus } from '../../services/study/loadflow';
 import { OptionalServicesNames } from '../utils/optional-services';
 import { useOptionalServiceStatus } from '../../hooks/use-optional-service-status';
 import { fetchNonEvacuatedEnergyStatus } from '../../services/study/non-evacuated-energy';
-import { useVoltageInitComputingStatus } from '../../hooks/use-voltage-init-computing-status';
 
 const loadFlowStatusInvalidations = ['loadflow_status', 'loadflow_failed'];
 
@@ -55,6 +56,10 @@ const oneBusShortCircuitAnalysisStatusInvalidations = [
 const dynamicSimulationStatusInvalidations = [
     'dynamicSimulation_status',
     'dynamicSimulation_failed',
+];
+const voltageInitStatusInvalidations = [
+    'voltageInit_status',
+    'voltageInit_failed',
 ];
 
 const loadFlowStatusCompletions = ['loadflowResult', 'loadflow_failed'];
@@ -83,6 +88,10 @@ const dynamicSimulationStatusCompletions = [
     'dynamicSimulationResult',
     'dynamicSimulation_failed',
 ];
+const voltageInitStatusCompletions = [
+    'voltageInitResult',
+    'voltageInit_failed',
+];
 
 // this hook loads all current computation status into redux then keeps them up to date according to notifications
 export const useAllComputingStatus = (
@@ -100,6 +109,9 @@ export const useAllComputingStatus = (
     );
     const dynamicSimulationAvailability = useOptionalServiceStatus(
         OptionalServicesNames.DynamicSimulation
+    );
+    const voltageInitAvailability = useOptionalServiceStatus(
+        OptionalServicesNames.VoltageInit
     );
     const shortCircuitAvailability = useOptionalServiceStatus(
         OptionalServicesNames.ShortCircuit
@@ -181,5 +193,14 @@ export const useAllComputingStatus = (
         dynamicSimulationAvailability
     );
 
-    useVoltageInitComputingStatus(studyUuid, currentNodeUuid);
+    useComputingStatus(
+        studyUuid,
+        currentNodeUuid,
+        fetchVoltageInitStatus,
+        voltageInitStatusInvalidations,
+        voltageInitStatusCompletions,
+        getVoltageInitRunningStatus,
+        ComputingType.VOLTAGE_INIT,
+        voltageInitAvailability
+    );
 };
